+++ contracts/test/RedeemableERC20Reentrant.sol
@@ -28,7 +28,7 @@ contract RedeemableERC20Reentrant is ReserveToken {
     function _beforeTokenTransfer(
         address sender_,
         address receiver_,
-        uint256 amount_
+        uint amount_
     ) internal virtual override {
         super._beforeTokenTransfer(sender_, receiver_, amount_);
         if (
